Coffee Shop Project
I wrote a simple Python program to simulate a coffee shop billing system using basic concepts like if, elif, and a while loop. The code allows the user to select items from a menu  and based on the choice, it adds the corresponding price to the total. The while loop keeps running until the user chooses to finish the order. After all selections, the program calculates and prints the total bill amount as a payment slip. This small project helped reinforce my understanding of decision-making and loops in Python.